Moving files out of settings folder.
